Implement TrueType/OpenType font support
This commit implements basic support for TrueType/OpenType fonts,
using FreeType to rasterize glyphs on the fly. This implementation
supports both static and variable fonts.

This is mainly intended to be a drop-in replacement for bitmap fonts,
with highly similar interfaces. Like bitmap fonts, the scalability
otherwise implied by the usage of TrueType/OpenType fonts is not
present, and text will blur at larger sizes. Unlike existing bitmap
fonts, the sizing is based on em box rather than line height, with
adjustments so that the text won't look too awful when font types are
mixed.